REFACTOR : 상세페이지 util  코드 분리
